+++ batch/batches/Storage/Periodic/KAsyncStoragePeriodicDeleteExe.php
@@ -1,0 +1,13 @@
+<?php
+/**
+ * Will run KAsyncStoragePeriodicDelete.class.php
+ * 
+ *
+ * @package Scheduler
+ * @subpackage Storage
+ */
+require_once(__DIR__ . "/../../../bootstrap.php");
+
+$instance = new KAsyncStoragePeriodicDelete();
+$instance->run(); 
+$instance->done();