+++ admin_console/forms/Partner/StorageConfiguration.php
@@ -149,6 +149,14 @@ class Form_Partner_StorageConfiguration extends Form_Partner_BaseStorageConfigur
 		));
 		$this->addElementToDisplayGroup('advanced', 'shouldExportThumbnails');
 
+		$this->addElement('checkbox', 'shouldExportCaptions', array(
+			'label'			=> "Should export caption assets",
+			'checked'		=> false,
+			'indicator'		=> 'dynamic',
+			'decorators' => array('ViewHelper', array('Label', array('placement' => 'append')), array('HtmlTag',  array('tag' => 'dt')))
+		));
+		$this->addElementToDisplayGroup('advanced', 'shouldExportCaptions');
+
 		$this->addPackagerConfigurationFields();
 	}
 
@@ -160,6 +168,7 @@ class Form_Partner_StorageConfiguration extends Form_Partner_BaseStorageConfigur
 	    parent::populateFromObject($object, $add_underscore);
 	    $this->setDefault('pathManagerParams', json_encode($object->pathManagerParams));
 		$this->setDefault('shouldExportThumbnails', $object->shouldExportThumbs);
+		$this->setDefault('shouldExportCaptions', $object->shouldExportCaptions);
 	}
 	
     public function getObject($objectType, array $properties, $add_underscore = true, $include_empty_fields = false)
@@ -192,6 +201,7 @@ class Form_Partner_StorageConfiguration extends Form_Partner_BaseStorageConfigur
 		
 		$object->pathManagerParams = json_decode($properties['pathManagerParams'], true);
 		$object->shouldExportThumbs = $properties['shouldExportThumbnails'];
+		$object->shouldExportCaptions = $properties['shouldExportCaptions'];
 		return $object;
 	}