+++ alpha/lib/model/StorageProfile.php
@@ -39,6 +39,7 @@ class StorageProfile extends BaseStorageProfile implements IBaseObject
 	const CUSTOM_DATA_MAPPED_PACKAGER_URL = 'mapped_packager_url';
 	const CUSTOM_DATA_EXPORT_PERIODICALLY = 'export_periodically';
 	const CUSTOM_DATA_EXCLUDED_FLAVOR_PARAMS_IDS = 'excluded_flavor_params_ids';
+	const CUSTOM_DATA_SHOULD_EXPORT_CAPTIONS ='should_export_captions';
 	/**
 	 * @var kStorageProfileScope
 	 */
@@ -187,6 +188,11 @@ class StorageProfile extends BaseStorageProfile implements IBaseObject
 
 		if ($flavorAsset instanceof thumbAsset)
 			return $this->getShouldExportThumbs();
+
+		if ($flavorAsset instanceof captionAsset)
+		{
+			return $this->getShouldExportCaptions();
+		}
 
 		if(!$this->isFlavorAssetConfiguredForExport($flavorAsset))
 		{
@@ -500,5 +506,15 @@ class StorageProfile extends BaseStorageProfile implements IBaseObject
 			}
 		}
 		return false;
+	}
+
+	public function getShouldExportCaptions()
+	{
+		return $this->getFromCustomData(self::CUSTOM_DATA_SHOULD_EXPORT_CAPTIONS,null, false);
+	}
+
+	public function setShouldExportCaptions($v)
+	{
+		$this->putInCustomData(self::CUSTOM_DATA_SHOULD_EXPORT_CAPTIONS, $v);
 	}
 }