+++ alpha/lib/model/assetPeer.php
@@ -795,7 +795,7 @@ class assetPeer extends BaseassetPeer implements IRelatedObjectPeer
 		return false;
 	}
 
-	public static function retrieveDscFlavorsByEntryIdAndStatus($entryId, array $statuses = array())
+	public static function retrieveDscFlavorsByEntryIdAndStatus($entryId, array $statuses = array(), $flavorTypes = array())
 	{
 		$c = new Criteria();
 		$c->add(assetPeer::ENTRY_ID, $entryId);
@@ -804,8 +804,6 @@ class assetPeer extends BaseassetPeer implements IRelatedObjectPeer
 		{
 			$c->add(assetPeer::STATUS, $statuses, Criteria::IN);
 		}
-
-		$flavorTypes = self::retrieveAllFlavorsTypes();
 		$c->add(assetPeer::TYPE, $flavorTypes, Criteria::IN);
 
 		$c->addDescendingOrderByColumn(assetPeer::INT_ID);