+++ api_v3/lib/types/storageProfile/KalturaStorageProfile.php
@@ -208,6 +208,11 @@ class KalturaStorageProfile extends KalturaObject implements IFilterable
 	 */
 	public $excludedFlavorParamsIds;
 
+	/**
+	 * @var bool
+	 */
+	public $shouldExportCaptions;
+
 	private static $map_between_objects = array
 	(
 		'id',
@@ -245,7 +250,8 @@ class KalturaStorageProfile extends KalturaObject implements IFilterable
 		'mappedPackagerUrl',
 		'regularPackagerUrl',
 		'exportPeriodically',
-		'excludedFlavorParamsIds'
+		'excludedFlavorParamsIds',
+		"shouldExportCaptions",
 	);
 	
 	/* (non-PHPdoc)